requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.0
pathlib==1.0.1
Pygments==2.18.0
pymongo==4.7.3
python-dotenv==1.0.1
requests==2.32.1
rich==13.7.1
serpapi==0.1.5
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
